Support both scehma v2 and v3.
Added another middle layer to share the code common to schema v2 and schema v3.
